Update README.md
The US Federal Constitution gives the President the power to amnesty various criminals. The amnesty is unilaterally determined by the president, and not only cannot be changed, even Congress and the courts have no right to review it. Ralph Stein, a professor at Pace University School of Law in New York State, pointed out that the president is not legally obliged to explain the reasons for the amnesty.

Trump's amnesty for Kushner and other people who have close interests with Trump may constitute a conflict of interest, that is, not out of public interest but to serve his personal interests. What proportion of the amnesty by  US presidents served their personal interests ?  What proportion of the amnesty served the public interest? This is an interesting entry point for examining the American system.

There are some difficulties to research such an object. First, it is difficult to obtain the relevant information of the amnesty object, so it is difficult to investigate whether the amnesty object is related to the president’s personal interests. Second, because of the high difficulty of obtaining information, the mining of useful information faces high costs. High financial support is needed. Third, various research funds lack the enthusiasm to support this kind of research that seems to negate the US system. Fourth, funds interested in supporting such research may be motivated to deny the US system, so the authoritarian countries’  funds more likely to give support , which involves the question of the "foreign agent" .

After thinking about this issue, I think it is possible to complete this research by such a model as "big data + AI + internet community collaboration". First, the large amount of information on the Internet—especially the information that can be searched through Google—provides a high possibility to sort out the relationship between the amnesty object and the president. Second, the open and free AI platform is low. The cost of parsing text from massive data provides computing power to match the interest relationship between the two. Third, the results of AI parsing may be biased, and the efforts of economists, political scientists, and computer scientists are integrated through community cooperation. This kind of bias may be corrected.. Fourth, if there are enough people involved in this kind of community cooperation, the tasks shared by each may not be too heavy, and it is possible to avoid project failure through voluntary means. Fifth, because it is a voluntary method, it can also be avoided The question of "foreign agent" guarantees the neutrality of the project.

I am an economist in China with few AI skill. I don’t know whether my thinking is feasible, so I established this project to call on interested economists, political scientists and computer scientists to try to cooperate.

美国联邦宪法给予总统特赦各种罪犯的权力。特赦由总统单方面定夺，不但不能更改，就连国会和法庭也无权审查。纽约州佩斯大学法学院教授拉尔夫.施泰因指出，总统在法律上没有义务解释特赦的理由。

川普关于库十纳等与川普有密切利益关系的人的特赦可能构成利益冲突，即不是出于公共利益而是为了服务于他的个人利益。美国历届总统的特赦到底在多大比例上是服务于总统的个人利益的？多大程度上是服务于公共利益的？这是考察美国制度的一个有趣切入点。

这种研究的难点在于：第一，特赦对象的相关信息难以获取，因此特赦对象是否与总统个人利益相关难以考察；第二，由于信息获取难度高，因此挖掘有用信息面临高昂成本，这种研究需要高昂的财务支持；第三，各种研究基金缺乏积极性支持这种似乎是否定美国制度的研究；第四，有兴趣支持这种研究的基金可能出于否定美国制度的动机，因此专制国家的基金可能支持这种研究，这涉及到“外国代理人”类型的质疑。

我思考这个问题后，认为在“大数据+AI+社区协作”模式下有可能完成这项研究。第一，互联网上大量的信息——特别是通过Google可以搜索到的信息——为整理出特赦对象与总统的利益关系提供了很高的可能性；第二，开放、免费的AI平台为低成本的从海量数据里解析文本对二者利益关系进行匹配提供了算力；第三，AI解析的结果可能有偏误，通过社区合作整合经济学家、政治学家与计算机科学家的努力，有可能纠正这种偏误；第四，如果这种社区合作参与的人足够多，各自分担的任务可能不是太重，通过志愿方式有可能避免项目夭折；第五，由于是志愿方式，还能避免“外国代理人”这种质疑，保证项目的中立性。

我不知道自己的思考是否可行，因此建立这个项目呼吁感兴趣的经济学家、政治学家与计算机科学家一起来尝试推进这项研究。